Added new classes Global, PathConfig, UserConfig and RuntimeConfig
Those four classes can be seen as a first draft of a replacement for
"engine" and "configManager". Instead of

  engine::hasGui()
  configManager::inst()->backgroundArtwork()
  configManager::inst()->dataDir()

one would write now:

  Global::runtimeConfig().hasGui()
  Global::userConfig().backgroundArtwork()
  Global::paths().dataDir()

Most of the members of configManager already have been copied to
PathConfig or UserConfig and can be used everywhere as soon as
the new preference dialog makes use of them.

All members of engine will be moved to Global.