PreferencesDialog: added ConfigUiAdaptor + Doxygen comments
Added a new internal helper class ConfigUiAdaptor which allows to link
an input widget and a configuration property. All we need in the
preferences dialog is to create one adaptor for each input widget+property,
e.g. a QCheckBox and a boolean property.

The macro ADD_USER_CONFIG_UI_ADAPTOR eases this even more. One only
needs to specify a property name.

Furthermore commented the PreferencesDialog class in Doxygen style.