AudioOutputContext::FifoBuffer: added lockless operation mode
Added an alternative implementation for the FIFO read and write
functions so they are lockless. This is realized via spinlocks and thus
consumes more CPU. The advantage however is less latency (when talking
about CPU cycles) and better realtime safety.

This new mode is used with all callback-based audio backends. This
should lead to less xruns with the JACK backend. However there's still
room for improvements.